kljaljfefjaéjdfléajfas

+++ src/Directions.java
@@ -4,7 +4,7 @@ import lejos.hardware.sensor.SensorMode;
 
 public class Directions {
 
-	public void direction(EV3LargeRegulatedMotor moteurDroit, EV3LargeRegulatedMotor moteurGauche, int[] valeurCouleur) {
+	public void direction(EV3LargeRegulatedMotor moteurDroit, EV3LargeRegulatedMotor moteurGauche, float[] valeurCouleur) {
